backend/requirements.txt
fastapi
uvicorn
motor
python-jose
httpx
python-dotenv
